chore(scap): move (some) udig-specific stuff into engine/udig
